os-refresh-config script to invoke config hooks
This orc script takes a list of heat config metadata under
the 'deployments' key and invokes a hook for each config item.

Hooks are stored in /var/lib/heat-config/hooks and the config group
name is used to specify which hook script to invoke.

A hook receives the config JSON on stdin and writes a JSON dict containing
outputs data to stdout. The orc script takes the outputs dict and signals
that back to heat using either a cfn signal post or a heatclient
resource-signal library call.

The following changes in this series contain hook implementations
and example templates.

Change-Id: Ia02005c28d550a195771dfacbba5fd71a39db709