Software config hook to provision using shell scripts
Deployment input values are mapped to shell environment variables.
Scripts can write output values to $heat_outputs_path.{output_name}
These output values will be signalled back to heat, making these
values available to the stack, and placing that deployment resource in a
COMPLETE state.

The example template demonstrates deploying a single config with more
that one deployment resource, each with different input values.

The only reason this is the first hook to be authored is because
a script is a "configuration tool" which requires no tooling to
set up.

Closes-Bug: #1279064

Change-Id: Ib297957ad0f89e34b7930ba97886116bff1e313c