Add env_file support for docker-compose hook
Added support for `env_file` which can be single file or a list
of files. Use special input 'env_files' for the env_files. These
files will be created by the hook relative to the compose project
directory.

if 'env_file' entries in 'docker-compose.yml' do not exist in the
`input_values` supplied, docker-compose will throw an error as it
can't find these files.

Change-Id: I0dbe80e8c8c56cc680f8ee86b4f34cf3d27533c4