Add script so that deployments won't get rerun
This script creates the necessary files under /var/run/heat-config so that the
next time configure.d/55-heat-config is executed via os-refresh-config, no
deployments that were already known are rerun.

This is a workaround for https://bugs.launchpad.net/heat-templates/+bug/1513220
in scenarios where /var/run/heat-config has already been lost due to a scenario
such as a system reboot.

This script is a manual workaround, and should only be executed on instances
where it's known that the deployments do not need to be reapplied.

Change-Id: I3ac4e280a39acb893a8ecc94712eb6265a1236d2
Partial-Bug: #1513220