Switch no hook message to an error
If the end host does not have the appropriate heat-agent hook installed,
the message for this is printed out as a warning. It's more likely that
this is an error condition if a user is configuring a stack to be
deployed with a specific tool and it is not installed when the stack is
run.  As a UX improvement, let's increase the level from warning to
error to increase visibility for the operator.

Change-Id: Ief87d526cc5941179cea3f8ffcac47c3dabaf71b
Related-Bug: #1651785