Add exit_codes option to docker-cmd hook
Some command we need to run with 'exec' or 'run' have custom
exit codes that are valid if the command completes sucessfully.
This patch adds an 'exit_codes' option which takes an array
of integers that are meant to represent the valid exit codes
for this hook. If for example 0 or 1 is meant to be considered
success you can specify:
 exit_codes: [0,1]

Previously an non-zero exit code would always fail this hook.

Change-Id: Ia6ca4b01982a0b33b26eca2a907d9d9f87c19922