Simplify legacy hieradata check to avoid false positive.
It seems that the current check for legacy hieradata can be fooled by
the way os-apply-config get its cached data.  This simplify the check
to avoid those false positives.

This a stable/ocata patch only as it's the transition between old and
new hiera mechanism.  For master, the stricter check is kept.

Change-Id: I53eddcb4e80f1256feb4285fdc7d25a2d70fb593
Closes-Bug: #1680006