Update requirements manually
The global requirments sync for this repo fails since it lists
packages that are not in the global list. Manually sync
requirements.

Failure log from requirements sync:
http://logs.openstack.org/d8/d89aad257d43fe02e42e44b2a0187dbd9f505c36/post/propose-update-requirements/15a7777/job-output.txt.gz#_2018-01-31_06_51_56_655451

Errors are:
'discover' is not in global-requirements.txt or blacklist.txt
'dpath' is not in global-requirements.txt or blacklist.txt
'salt' is not in global-requirements.txt or blacklist.txt
'yamllint' is not in global-requirements.txt or blacklist.txt

Change-Id: Ifbd174d568b0cae5225faa6e25b1bc5fc4774e6e