import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

This patch also adds limiting dependencies using the upper-constraints
file in tox.ini, as this is needed to unblock the gate.

Change-Id: Ib2ef3b3d0d0733842abad97898114c3dbdab3b7a
Co-Authored-By: Zane Bitter <zbitter@redhat.com>
Story: #2002586
Task: #24298