[Ansible] Look for inventory in SoftwareConfig options
Currently, the only way to use a custom inventory is to change the
HEAT_ANSIBLE_INVENTORY environment variable. This becomes a problem when
the inventory has to be changed between playbook runs from two
sequential SoftwareDeployments, as environment variables won't be
reloaded.

This change allows to specify the location of the inventory in the
"options" map for each SoftwareConfig.

Change-Id: I6709213240119e999d4f36e2938b0de9311d2556